cluster/pyproject.toml
[project]
name = "cluster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "humanfriendly>=10.0",
    "pydantic-settings>=2.9.1",
    "pympler>=1.1",
    "pytest>=8.3.5",
    "random-word>=1.0.13",
]

[tool.ruff]
line-length = 80

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
markers = ["integtest", "long"]
